pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "markdown_convert"
version = "1.2.53"
description = "Convert Markdown files to PDF from your command line."
authors = [
    { name = "Julio Cabria", email = "juliocabria@tutanota.com" },
]
license = { text = "GPL-2.0-only" }
readme = "README.md"
requires-python = ">=3.11,<3.15"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Operating System :: OS Independent",
]
dependencies = [
    "argsdict==1.0.0",
    "markdown2>=2.4.13,<3",
    "pygments>=2.17.2,<3",
    "latex2mathml>=3.78.1",
    "playwright>=1.57.0",
    "beautifulsoup4>=4.14.3",
    "install-playwright>=1.0.0",
    "vl-convert-python>=1.9.0.post1",
    "ruamel-yaml>=0.19.1",
]

[project.urls]
homepage = "https://github.com/Julynx/markdown_convert"

[project.scripts]
markdown-convert = "markdown_convert.__main__:main"

[tool.hatch.build.targets.sdist]
include = [
    "markdown_convert/__init__.py",
    "markdown_convert/__main__.py",
    "markdown_convert/default.css",
    "markdown_convert/code.css",
    "markdown_convert/modules/__init__.py",
    "markdown_convert/modules/constants.py",
    "markdown_convert/modules/convert.py",
    "markdown_convert/modules/transform.py",
    "markdown_convert/modules/extras.py",
    "markdown_convert/modules/resources.py",
    "markdown_convert/modules/utils.py",
    "markdown_convert/modules/validate.py",
    "markdown_convert/modules/autoinstall.py",
    "markdown_convert/modules/overrides.py",
]

[tool.hatch.build.targets.wheel]
packages = ["markdown_convert"]
include = [
    "markdown_convert/__init__.py",
    "markdown_convert/__main__.py",
    "markdown_convert/default.css",
    "markdown_convert/code.css",
    "markdown_convert/modules/__init__.py",
    "markdown_convert/modules/constants.py",
    "markdown_convert/modules/convert.py",
    "markdown_convert/modules/transform.py",
    "markdown_convert/modules/extras.py",
    "markdown_convert/modules/resources.py",
    "markdown_convert/modules/utils.py",
    "markdown_convert/modules/validate.py",
    "markdown_convert/modules/autoinstall.py",
    "markdown_convert/modules/overrides.py",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "pylint>=3.3.9",
    "ruff>=0.14.11",
    "yina>=0.1.6",
]
